requirements.txt
Django>=4.2.0,<5.0.0
django-filter>=23.5
djangorestframework>=3.14.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.9
whitenoise>=6.6.0
python-dotenv>=1.0.0
dj-database-url>=2.1.0 